FA-91 loginCompose module
