--- .idea/modules/createAccountCompose/ForzenBook.createAccountCompose.androidTest.iml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="TestModuleProperties" production-module="ForzenBook.createAccountCompose.main" />
-</module>
