--- .idea/modules/loginCompose/ForzenBook.loginCompose.unitTest.iml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="TestModuleProperties" production-module="ForzenBook.loginCompose.main" />
-</module>
