+++ loginCompose/src/main/res/values/strings.xml
@@ -1,0 +1,19 @@
+<resources>
+
+    <string name="login_title">forzenbook</string>
+    <string name="login_image_description">Forzen Lion Logo</string>
+    <string name="login_button_text">Login</string>
+    <string name="login_email_hint">Email</string>
+    <string name="login_email_error">Please Enter a Valid Email</string>
+    <string name="login_code_hint">Code</string>
+    <string name="login_code_error">Please Enter Code</string>
+    <string name="login_char_length_limit">Max Character Limit Reached</string>
+    <string name="login_invalid_credentials">Invalid Credential</string>
+    <string name="login_get_code_text">Send Code</string>
+    <string name="login_create_account">Create Account</string>
+    <string name="required_fields_error">All Fields Required</string>
+    <string name="error_no_internet_connection">No Internet Connection</string>
+    <string name="error_dismiss">Dismiss</string>
+    <string name="error_connect_and_try_again">Check Connection and Try Again</string>
+
+</resources>
