+++ settings.gradle
@@ -19,3 +19,4 @@ include ':loginCore'
 include ':createAccountCore'
 include ':composeCore'
 include ':uiCore'
+include ':loginCompose'
